+++ assets/js/scripts.js
@@ -1,12 +1,16 @@
 window.onload = function() {
-    var menuBtn = document.querySelector("header nav span"),
-        menu = document.querySelector("header nav ul");
-
+    let menuBtn = document.querySelector("header nav span");
+    let menu = document.querySelector("header nav ul");
+    
     menuBtn.onclick = function() {
-        if (menu.classList.contains("show")) {
-            menu.classList.remove("show");
-        } else {
-            menu.classList.add("show");
-        }
+        menu.classList.toggle("show");
     };
+    
+    if (window.location.pathname == "/admin/users.php" || window.location.pathname == "/admin/games.php" || window.location.pathname == "/admin/reports.php") {
+        new DataTable(".dataTable");
+    }
+
+    if (window.location.pathname == "/report.php") {
+        NiceSelect.bind(document.querySelector(".select2"), {searchable: true});
+    }
 };